Convert Ranking and Team models to use Bolt DB.

+++ field/arena_notifiers.go
@@ -85,7 +85,7 @@ func (arena *Arena) generateArenaStatusMessage() interface{} {
 	}
 
 	return &struct {
-		MatchId          int64
+		MatchId          int
 		AllianceStations map[string]*AllianceStation
 		TeamWifiStatuses map[string]network.TeamWifiStatus
 		MatchState
