+++ field/arena_test.go
@@ -388,13 +388,13 @@ func TestLoadNextMatch(t *testing.T) {
 	arena.Database.CreateMatch(&qualificationMatch2)
 
 	// Test match should be followed by another, empty test match.
-	assert.Equal(t, int64(0), arena.CurrentMatch.Id)
+	assert.Equal(t, 0, arena.CurrentMatch.Id)
 	err := arena.SubstituteTeam(1114, "R1")
 	assert.Nil(t, err)
 	arena.CurrentMatch.Status = model.TieMatch
 	err = arena.LoadNextMatch()
 	assert.Nil(t, err)
-	assert.Equal(t, int64(0), arena.CurrentMatch.Id)
+	assert.Equal(t, 0, arena.CurrentMatch.Id)
 	assert.Equal(t, 0, arena.CurrentMatch.Red1)
 	assert.Equal(t, false, arena.CurrentMatch.IsComplete())
 
@@ -413,7 +413,7 @@ func TestLoadNextMatch(t *testing.T) {
 	arena.Database.UpdateMatch(&practiceMatch3)
 	err = arena.LoadNextMatch()
 	assert.Nil(t, err)
-	assert.Equal(t, int64(0), arena.CurrentMatch.Id)
+	assert.Equal(t, 0, arena.CurrentMatch.Id)
 	assert.Equal(t, "test", arena.CurrentMatch.Type)
 
 	err = arena.LoadMatch(&qualificationMatch1)
